+++ c_port/readRawText.h
@@ -5,6 +5,6 @@
 #define READRAW_H
 
 //readRawText.c function declaration
-float *readRawText(char *data_dir, char *config_prefix, char *config_filename, int data_row, int data_col, float buff[][data_col]);	
-
+double *readRawText(char *data_dir, char *config_prefix, char *config_filename,
+					int data_row, int data_col, double buff[][data_col]);
 #endif					// CONFIG_H
